refactor(core): move the generic EventChannel safeCancel guard out of obd2 into core/utils (#3134)
The cancel-race guard is fully generic (used by core/location and the
widget feature, not just OBD2), so core importing it from
features/consumption was a layering inversion. Move it to
lib/core/utils/, update the six importers + the test, drop the
consumption api.dart export, and lock the boundary-baseline
decrements in (core->consumption 8->6, widget->consumption pair gone).
